server/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Session==0.5.0
Flask-Security-Too==5.3.2
Flask-Admin==1.6.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AWS
boto3==1.29.3
botocore==1.32.3

# Authentication and Security
PyJWT==2.8.0
pyotp==2.9.0
bcrypt==4.0.1
cryptography==41.0.5

# File handling
Pillow==10.1.0
python-magic==0.4.27
Werkzeug==3.0.1

# Notifications
twilio==8.10.0
sendgrid==6.10.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
validators==0.22.0
phonenumbers==8.13.24
pytz==2023.3.post1
python-dateutil==2.8.2
email-validator==2.1.0.post1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
factory-boy==3.3.0
faker==20.0.3

# Development
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1

# Production
gunicorn==21.2.0
redis==5.0.1
sentry-sdk==1.32.0
newrelic==9.1.0
